Move all project core files into a src folder
