+++ {{cookiecutter.project_dirname}}/__tests__/components/Navbar.test.tsx
@@ -2,7 +2,7 @@ import { expect } from '@jest/globals'
 import { screen } from '@testing-library/react'
 import React from 'react'
 
-import { Navbar } from '@/components/Navbar'
+import { Navbar } from '@/src/components/Navbar'
 import { renderWithWrappers } from '@/__tests__/functions'
 
 const setup = () => renderWithWrappers(<Navbar />)
