+++ {{cookiecutter.project_dirname}}/__tests__/components/commons/Button.test.tsx
@@ -2,7 +2,7 @@ import { expect } from '@jest/globals'
 import { screen } from '@testing-library/react'
 import React from 'react'
 
-import { Button } from '@/components/commons/Button'
+import { Button } from '@/src/components/commons/Button'
 import { renderWithWrappers } from '@/__tests__/functions'
 
 const solidSetup = () => renderWithWrappers(<Button />)
