+++ {{cookiecutter.project_dirname}}/cypress/e2e/Health/Health.cy.ts
@@ -1,7 +1,7 @@
 context('Health check endpoint', () => {
   describe('Service health endpoint', () => {
     it('returns a 204 status', () => {
-      cy.request('/{{ cookiecutter.service_slug }}/health').then(response => {
+      cy.request('/frontend/health').then(response => {
         expect(response.status).to.equal(204)
         expect(response.body).to.be.empty
       })
