+++ {{cookiecutter.project_dirname}}/src/models/Utils.ts
