+++ {{cookiecutter.project_dirname}}/src/styles/GlobalStyle.tsx
