+++ {{cookiecutter.project_dirname}}/src/utils/api/index.ts
@@ -1,0 +1,1 @@
+export * from '@/src/utils/api/auth'
