--- {{cookiecutter.project_dirname}}/utils/api/index.ts
@@ -1,1 +1,0 @@
-export * from '@/utils/api/auth'
